+++ app/serializers/coaching/person_serializer.rb
@@ -1,0 +1,5 @@
+class Coaching::PersonSerializer < PersonSerializer
+  def person_exhibit
+    @exhibit ||= Coaching::PersonExhibit.new(object, nil)
+  end
+end